4.15-4.22

+++ osa4/blogilista/controllers/blogs.js
@@ -1,8 +1,11 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs.map(blog => blog.toJSON()))
 })
 
@@ -17,24 +20,47 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
-    
+            
     if(!body.title || !body.url) {
         return response.status(400).end()
     }
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if(!request.token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
 
     const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes
+        likes: body.likes,
+        user: user._id
     })
     
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     response.status(201).json(savedBlog.toJSON())
-        
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const blog = await Blog.findById(request.params.id)
+    if(!blog) {
+        return response.status(404).end()    
+    }
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if(!request.token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+    
+    if(!blog.user || (blog.user.toString() !== user._id.toString())) {
+        return response.status(401).end()
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
